Refcator: rename utilities to helpers
