+++ src/components/TransactionSigner.js
@@ -19,15 +19,15 @@ import OptionsTablePair from "./OptionsTable/Pair";
 import SecretKeyPicker from "./FormComponents/SecretKeyPicker";
 import MultiPicker from "./FormComponents/MultiPicker";
 import BipPathPicker from "./FormComponents/BipPathPicker";
-import { txPostLink, xdrViewer, feeBumpTxLink } from "../utilities/linkBuilder";
+import { txPostLink, xdrViewer, feeBumpTxLink } from "../helpers/linkBuilder";
 import HelpMark from "./HelpMark";
-import clickToSelect from "../utilities/clickToSelect";
-import scrollOnAnchorOpen from "../utilities/scrollOnAnchorOpen";
-import extrapolateFromXdr from "../utilities/extrapolateFromXdr";
-import validateTxXdr from "../utilities/validateTxXdr";
+import clickToSelect from "../helpers/clickToSelect";
+import scrollOnAnchorOpen from "../helpers/scrollOnAnchorOpen";
+import extrapolateFromXdr from "../helpers/extrapolateFromXdr";
+import validateTxXdr from "../helpers/validateTxXdr";
 import NETWORK from "../constants/network";
-import Libify from "../utilities/Libify";
-import { addEventHandler } from "../utilities/metrics";
+import Libify from "../helpers/Libify";
+import { addEventHandler } from "../helpers/metrics";
 import transactionSignerMetrics from "../metricsHandlers/transactionSigner";
 
 const { signTransaction } = Libify;
