+++ src/components/TxBuilderResult.js
@@ -3,9 +3,9 @@ import { connect } from "react-redux";
 import { isConnected } from "@stellar/freighter-api";
 import reduce from "lodash/reduce";
 import { EasySelect } from "./EasySelect";
-import Libify from "../utilities/Libify";
-import { txSignerLink, xdrViewer } from "../utilities/linkBuilder";
-import scrollOnAnchorOpen from "../utilities/scrollOnAnchorOpen";
+import Libify from "../helpers/Libify";
+import { txSignerLink, xdrViewer } from "../helpers/linkBuilder";
+import scrollOnAnchorOpen from "../helpers/scrollOnAnchorOpen";
 import TX_TYPES from "../constants/transaction_types";
 
 class TxBuilderResult extends React.Component {
