+++ src/helpers/clickToSelect.js
