+++ src/helpers/linkHighlighterRules.js
