+++ src/helpers/sanitizeHtml.js
