+++ src/helpers/scrollOnAnchorOpen.js
