+++ src/helpers/trezorTransformTransaction.js
