+++ src/helpers/validateBase64.js
