+++ src/helpers/validateTxXdr.js
