+++ src/reducers/transactionSigner.js
@@ -15,7 +15,7 @@ import {
   FREIGHTER_WALLET_SIGN_SUCCESS,
 } from "../actions/transactionSigner";
 import { LOAD_STATE } from "../actions/routing";
-import validateTxXdr from "../utilities/validateTxXdr";
+import validateTxXdr from "../helpers/validateTxXdr";
 import SLUG from "../constants/slug";
 
 const transactionSigner = combineReducers({
